Refactor: ./domain/** -> ./domain/common/**
