+++ src/main/java/rooftopgreenlight/urbanisland/api/service/FileService.java
@@ -6,7 +6,6 @@ import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.querydsl.core.types.Predicate;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -14,8 +13,8 @@ import rooftopgreenlight.urbanisland.api.common.exception.FileIOException;
 import rooftopgreenlight.urbanisland.api.common.properties.AwsS3Properties;
 import rooftopgreenlight.urbanisland.api.common.properties.ProfileDirProperties;
 import rooftopgreenlight.urbanisland.api.controller.dto.FileResponse;
-import rooftopgreenlight.urbanisland.domain.exception.NotFoundMemberException;
-import rooftopgreenlight.urbanisland.domain.exception.NotFoundProfileException;
+import rooftopgreenlight.urbanisland.domain.common.exception.NotFoundMemberException;
+import rooftopgreenlight.urbanisland.domain.common.exception.NotFoundProfileException;
 import rooftopgreenlight.urbanisland.domain.file.entity.Profile;
 import rooftopgreenlight.urbanisland.domain.file.entity.QProfile;
 import rooftopgreenlight.urbanisland.domain.file.repository.ProfileRepository;
