+++ src/main/java/rooftopgreenlight/urbanisland/domain/member/entity/Member.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import rooftopgreenlight.urbanisland.domain.BaseEntity;
+import rooftopgreenlight.urbanisland.domain.common.BaseEntity;
 import rooftopgreenlight.urbanisland.domain.file.entity.Profile;
 
 import javax.persistence.*;
